+++ 1.21.1/Common/src/main/java/fuzs/puzzleslib/impl/capability/ClientboundEntityCapabilityMessage.java
@@ -1,6 +1,8 @@
 package fuzs.puzzleslib.impl.capability;
 
+import com.mojang.serialization.DataResult;
 import fuzs.puzzleslib.api.capability.v3.CapabilityController;
+import fuzs.puzzleslib.api.capability.v3.data.CapabilityComponent;
 import fuzs.puzzleslib.api.capability.v3.data.CapabilityKey;
 import fuzs.puzzleslib.api.network.v3.ClientMessageListener;
 import fuzs.puzzleslib.api.network.v3.ClientboundMessage;
@@ -8,11 +10,15 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.NbtOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
 
-public record ClientboundEntityCapabilityMessage(ResourceLocation identifier, int entityId, CompoundTag tag) implements ClientboundMessage<ClientboundEntityCapabilityMessage> {
+public record ClientboundEntityCapabilityMessage(ResourceLocation identifier,
+                                                 int entityId,
+                                                 CompoundTag tag) implements ClientboundMessage<ClientboundEntityCapabilityMessage> {
 
     @Override
     public ClientMessageListener<ClientboundEntityCapabilityMessage> getHandler() {
@@ -23,8 +29,20 @@ public record ClientboundEntityCapabilityMessage(ResourceLocation identifier, in
                 Entity entity = level.getEntity(message.entityId);
                 if (entity != null) {
                     CapabilityKey<?, ?> capabilityKey = CapabilityController.get(message.identifier);
-                    capabilityKey.getIfProvided(entity).ifPresent(capabilityComponent -> capabilityComponent.read(message.tag, handler.registryAccess()));
+                    if (capabilityKey.isProvidedBy(entity)) {
+                        this.setCapabilityComponent(message.tag, entity, capabilityKey, handler.registryAccess());
+                    }
                 }
+            }
+
+            @SuppressWarnings("unchecked")
+            private <T extends Entity, C extends CapabilityComponent<T>> void setCapabilityComponent(CompoundTag tag, Entity entity, CapabilityKey<?, ?> capabilityKey, HolderLookup.Provider registries) {
+                ((DataResult<C>) capabilityKey.codec()
+                        .parse(registries.createSerializationContext(NbtOps.INSTANCE),
+                                tag
+                        )).ifSuccess((C capabilityComponent) -> {
+                    ((CapabilityKey<T, C>) capabilityKey).set((T) entity, capabilityComponent);
+                });
             }
         };
     }