+++ 1.21.1/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/core/NeoForgeFactories.java
@@ -7,9 +7,11 @@ import fuzs.puzzleslib.api.init.v3.GameRulesFactory;
 import fuzs.puzzleslib.api.init.v3.registry.RegistryFactory;
 import fuzs.puzzleslib.api.item.v2.ToolTypeHelper;
 import fuzs.puzzleslib.api.item.v2.crafting.CombinedIngredients;
+import fuzs.puzzleslib.impl.attachment.DataAttachmentRegistryImpl;
 import fuzs.puzzleslib.impl.core.CommonFactories;
 import fuzs.puzzleslib.impl.core.ModContext;
 import fuzs.puzzleslib.impl.core.ProxyImpl;
+import fuzs.puzzleslib.neoforge.impl.attachment.NeoForgeDataAttachmentRegistryImpl;
 import fuzs.puzzleslib.neoforge.impl.data.NeoForgeTagAppender;
 import fuzs.puzzleslib.neoforge.impl.event.NeoForgeEventInvokerRegistryImpl;
 import fuzs.puzzleslib.neoforge.impl.init.NeoForgeGameRulesFactory;
@@ -79,5 +81,10 @@ public final class NeoForgeFactories implements CommonFactories {
     @Override
     public <T> AbstractTagAppender<T> getTagAppender(TagBuilder tagBuilder, @Nullable Function<T, ResourceKey<T>> keyExtractor) {
         return new NeoForgeTagAppender<>(tagBuilder, keyExtractor);
+    }
+
+    @Override
+    public DataAttachmentRegistryImpl getDataAttachmentRegistry() {
+        return new NeoForgeDataAttachmentRegistryImpl();
     }
 }