+++ 1.21.1/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.5-1.21.1] - 2024-09-16
+### Added
+- Add new data attachment api
+### Changed
+- Revert codec support for capabilities, use attachments for that
+- Rename `PlayerSet::notify` to `PlayerSet::broadcast`
+
 ## [v21.1.4-1.21.1] - 2024-09-14
 ### Changed
 - Support codecs for capabilities